debugging why some dictionaries have sets for keys
